+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added `getAccount` query on OrganizationDetailsSellers component to get the account data
+
 ## [1.34.1] - 2024-08-21
 
 ### Fixed